fix tests

+++ Makefile
@@ -10,7 +10,7 @@ f fmt:
 	terraform fmt
 
 setup:
-	for i in tests/*; do ln -vfs $$i; done
+	[ -d tests ] && for i in tests/*; do ln -vfs $$i; done || true
 
 clean:
-	for i in tests/*; do rm -v $${i#tests/}; done
+	[ -d tests ] && for i in tests/*; do rm -v $${i#tests/}; done || true
